Play Services screen auto update
Demo mode (opened from DEV menu)
- always open Play Services screen
- always open Google Play

Normal mode
- whenever is the screen refreshed (eg. after the return from Google Play) is current version of Play Services checked
  * if the app is updated, screen is closed

- if Play Services are updated and screen is still shown, back button and update button redirect the user to previous screen
- if Play services are outdated back button close the app and update button open Google Play
