+++ app/src/main/res/navigation/nav_graph.xml
@@ -282,7 +282,10 @@
             android:name="fullscreen"
             android:defaultValue="true"
             app:argType="boolean" />
-
+        <argument
+            android:name="demo"
+            android:defaultValue="false"
+            app:argType="boolean" />
     </fragment>
     <fragment
         android:id="@+id/nav_dashboard_cards"
